HasEvent mark next/prev months
For consistence mark the next and prev month calendar days (visible in current month) with events. Now it is done only only during the initiation of calendar widget, the reason of modification is to have the same functionality when change in widget the view to next/prev month. Also this change can be applied with change in css file and adding another color for days .out-of-month.hasEvent for example add line like that 
.calendar-widget-table tbody .out-of-month.hasEvent {background: #33CC99;color:#fff;}