Update CalendarEvent.php
Designer here. Noticed other dates doesn't return anything if no date string is set in URL. Appears function doesn't return anything if that is the case.

This change solves it. Not sure if you had it your way for a reason, but thought maybe it's worth looking at as I needed this for a project.